# contracts/openapi.yaml
components:
  schemas:
...
    DatasetImportBindingRequest:
      additionalProperties: false
      properties:
        connection_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Connection Id
        role_assignments:
          items:
            $ref: '#/components/schemas/ImportRoleAssignmentRequest'
          title: Role Assignments
          type: array
        selected_object_ids:
          items:
            type: string
          title: Selected Object Ids
          type: array
        source_type:
          title: Source Type
          type: string
        upload_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Upload Id
      required:
      - source_type
      title: DatasetImportBindingRequest
      type: object
    DatasetImportBindingResponse:
      additionalProperties: false
      properties:
        connection_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Connection Id
        current_receipt_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Current Receipt Id
        role_assignments:
          items:
            $ref: '#/components/schemas/ImportRoleAssignmentRequest'
          title: Role Assignments
          type: array
        selected_object_ids:
          items:
            type: string
          title: Selected Object Ids
          type: array
        source_type:
          title: Source Type
          type: string
        updated_at:
          format: date-time
          title: Updated At
          type: string
        upload_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Upload Id
      required:
      - source_type
      - selected_object_ids
      - role_assignments
      - updated_at
      title: DatasetImportBindingResponse
      type: object
...
    DatasetResponse:
      additionalProperties: false
      properties:
        created_at:
          format: date-time
          title: Created At
          type: string
        dataset_id:
          title: Dataset Id
          type: string
        name:
          title: Name
          type: string
        status:
          title: Status
          type: string
        updated_at:
          format: date-time
          title: Updated At
          type: string
        workspace_id:
          title: Workspace Id
          type: string
      required:
      - dataset_id
      - workspace_id
      - name
      - status
      - created_at
      - updated_at
      title: DatasetResponse
      type: object
...
    GovernanceCheckResponse:
      additionalProperties: false
      properties:
        code:
          title: Code
          type: string
        evidence:
          additionalProperties:
            type: string
          title: Evidence
          type: object
        message:
          title: Message
          type: string
        passed:
          title: Passed
          type: boolean
        severity:
          title: Severity
          type: string
      required:
      - code
      - severity
      - passed
      - message
      title: GovernanceCheckResponse
      type: object
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    ImportReadinessReceiptResponse:
      additionalProperties: false
      properties:
        attention_count:
          title: Attention Count
          type: integer
        check_results:
          items:
            $ref: '#/components/schemas/GovernanceCheckResponse'
          title: Check Results
          type: array
        contract_version:
          title: Contract Version
          type: string
        dataset_id:
          title: Dataset Id
          type: string
        error_count:
          title: Error Count
          type: integer
        manifest_fingerprint:
          title: Manifest Fingerprint
          type: string
        receipt_id:
          title: Receipt Id
          type: string
        role_assignment_count:
          anyOf:
          - type: integer
          - type: 'null'
          title: Role Assignment Count
        selected_object_count:
          anyOf:
          - type: integer
          - type: 'null'
          title: Selected Object Count
        source_type:
          title: Source Type
          type: string
        status:
          title: Status
          type: string
        superseded:
          default: false
          title: Superseded
          type: boolean
        tenant_id:
          title: Tenant Id
          type: string
        verified_at:
          format: date-time
          title: Verified At
          type: string
        workspace_id:
          title: Workspace Id
          type: string
      required:
      - receipt_id
      - contract_version
      - tenant_id
      - workspace_id
      - dataset_id
      - source_type
      - status
      - check_results
      - error_count
      - attention_count
      - manifest_fingerprint
      - verified_at
      title: ImportReadinessReceiptResponse
      type: object
    ImportRoleAssignmentRequest:
      additionalProperties: false
      properties:
        object_id:
          title: Object Id
          type: string
        provider:
          title: Provider
          type: string
        role:
          title: Role
          type: string
      required:
      - object_id
      - role
      - provider
      title: ImportRoleAssignmentRequest
      type: object
...
    PublishReadinessReceiptResponse:
      additionalProperties: false
      properties:
        check_results:
          items:
            $ref: '#/components/schemas/GovernanceCheckResponse'
          title: Check Results
          type: array
        contract_fingerprint:
          title: Contract Fingerprint
          type: string
        contract_version:
          title: Contract Version
          type: string
        dataset_id:
          title: Dataset Id
          type: string
        destination_summaries:
          items:
            type: string
          title: Destination Summaries
          type: array
        model_ready_fingerprint:
          anyOf:
          - type: string
          - type: 'null'
          title: Model Ready Fingerprint
        published:
          default: false
          title: Published
          type: boolean
        receipt_id:
          title: Receipt Id
          type: string
        run_id:
          title: Run Id
          type: string
        status:
          title: Status
          type: string
        tenant_id:
          title: Tenant Id
          type: string
        verified_at:
          format: date-time
          title: Verified At
          type: string
        workspace_id:
          title: Workspace Id
          type: string
      required:
      - receipt_id
      - contract_version
      - tenant_id
      - workspace_id
      - dataset_id
      - run_id
      - status
      - destination_summaries
      - check_results
      - contract_fingerprint
      - verified_at
      title: PublishReadinessReceiptResponse
      type: object
...
    ValidationError:
      properties:
        ctx:
          title: Context
          type: object
        input:
          title: Input
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
...
  securitySchemes:
    HTTPBearer:
      bearerFormat: JWT
      description: Clerk session token forwarded by the Next.js BFF. prem3-api verifies the session token. Tenant IDs are
        never accepted from the client.
      scheme: bearer
      type: http
info:
  description: Presentation-safe Project, Dataset, upload, Evaluation, catalog, billing, Google connection, and import/publish
    governance contracts. Clerk session tokens are verified when the identity provider is configured. Creating an Evaluation
    returns 202 Accepted for resource creation only; durable ADK dispatch is not started from this HTTP boundary. Tenant identity
    is never accepted from the client. IMPORT_READY, MODEL_READY, and PUBLISH_READY are distinct deterministic states.
  summary: PreM3 authenticated product API
  title: prem3-api
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /v1/workspaces/{workspace_id}/datasets/{dataset_id}:
    get:
      operationId: getDataset
      parameters:
      - in: path
        name: dataset_id
        required: true
        schema:
          title: Dataset Id
          type: string
      - in: path
        name: workspace_id
        required: true
        schema:
          title: Workspace Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatasetResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Get Dataset
      tags:
      - datasets
...
  /v1/workspaces/{workspace_id}/datasets/{dataset_id}/evaluations/{run_id}/publish-readiness:
    post:
      operationId: evaluatePublishReadiness
      parameters:
      - in: path
        name: run_id
        required: true
        schema:
          title: Run Id
          type: string
      - in: path
        name: dataset_id
        required: true
        schema:
          title: Dataset Id
          type: string
      - in: path
        name: workspace_id
        required: true
        schema:
          title: Workspace Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublishReadinessReceiptResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Evaluate Publish Readiness Route
      tags:
      - import-governance
  /v1/workspaces/{workspace_id}/datasets/{dataset_id}/import-binding:
    get:
      operationId: getDatasetImportBinding
      parameters:
      - in: path
        name: dataset_id
        required: true
        schema:
          title: Dataset Id
          type: string
      - in: path
        name: workspace_id
        required: true
        schema:
          title: Workspace Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatasetImportBindingResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Get Import Binding
      tags:
      - import-governance
    put:
      operationId: putDatasetImportBinding
      parameters:
      - in: path
        name: dataset_id
        required: true
        schema:
          title: Dataset Id
          type: string
      - in: path
        name: workspace_id
        required: true
        schema:
          title: Workspace Id
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DatasetImportBindingRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatasetImportBindingResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Put Import Binding
      tags:
      - import-governance
  /v1/workspaces/{workspace_id}/datasets/{dataset_id}/import-readiness:
    get:
      operationId: getDatasetImportReadiness
      parameters:
      - in: path
        name: dataset_id
        required: true
        schema:
          title: Dataset Id
          type: string
      - in: path
        name: workspace_id
        required: true
        schema:
          title: Workspace Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportReadinessReceiptResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Get Import Readiness
      tags:
      - import-governance
    post:
      operationId: evaluateDatasetImportReadiness
      parameters:
      - in: path
        name: dataset_id
        required: true
        schema:
          title: Dataset Id
          type: string
      - in: path
        name: workspace_id
        required: true
        schema:
          title: Workspace Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportReadinessReceiptResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Evaluate Import Readiness Route
      tags:
      - import-governance
...